move function to query-feature.js
